fw-4130, split factories into multiple files
