+++ firstvoices/backend/tests/factories/__init__.py
@@ -1,3 +1,11 @@
-from .base import *  # noqa F401
-from .dictionary import *  # noqa F401
-from .media import *  # noqa F401
+from .access import *  # noqa F401, F403  # isort:skip
+from .app_factories import *  # noqa F401, F403  # isort:skip
+from .category_factories import *  # noqa F401, F403 # isort:skip
+from .site_info_factories import *  # noqa F401, F403 # isort:skip
+from .media_factories import *  # noqa F401, F403 # isort:skip
+
+from .character_factories import *  # noqa F401, F403 # isort:skip
+from .dictionary_entry import *  # noqa F401, F403 # isort:skip
+from .dictionary_factories import *  # noqa F401, F403 # isort:skip
+
+from .generic import *  # noqa F401, F403 # isort:skip
