+++ firstvoices/backend/tests/factories/dictionary_factories.py
@@ -3,11 +3,13 @@ from django.utils.timezone import datetime
 from factory.django import DjangoModelFactory
 
 from backend.models import PartOfSpeech, dictionary
-from backend.tests import factories
+from backend.tests.factories.access import SiteFactory, UserFactory
+from backend.tests.factories.character_factories import CharacterFactory
+from backend.tests.factories.dictionary_entry import DictionaryEntryFactory
 
 
 class DictionaryModelFactory(DjangoModelFactory):
-    dictionary_entry = factory.SubFactory(factories.DictionaryEntryFactory)
+    dictionary_entry = factory.SubFactory(DictionaryEntryFactory)
 
 
 class AcknowledgementFactory(DictionaryModelFactory):
@@ -39,8 +41,8 @@ class DictionaryEntryLinkFactory(DjangoModelFactory):
     class Meta:
         model = dictionary.DictionaryEntryLink
 
-    from_dictionary_entry = factory.SubFactory(factories.DictionaryEntryFactory)
-    to_dictionary_entry = factory.SubFactory(factories.DictionaryEntryFactory)
+    from_dictionary_entry = factory.SubFactory(DictionaryEntryFactory)
+    to_dictionary_entry = factory.SubFactory(DictionaryEntryFactory)
 
 
 class CategoryFactory(DjangoModelFactory):
@@ -48,16 +50,16 @@ class CategoryFactory(DjangoModelFactory):
         model = dictionary.Category
 
     title = factory.Sequence(lambda n: "Category %03d" % n)
-    site = factory.SubFactory(factories.SiteFactory)
-    created_by = factory.SubFactory(factories.UserFactory)
-    last_modified_by = factory.SubFactory(factories.UserFactory)
+    site = factory.SubFactory(SiteFactory)
+    created_by = factory.SubFactory(UserFactory)
+    last_modified_by = factory.SubFactory(UserFactory)
 
 
 class DictionaryEntryCategoryFactory(DjangoModelFactory):
     class Meta:
         model = dictionary.DictionaryEntryCategory
 
-    dictionary_entry = factory.SubFactory(factories.DictionaryEntryFactory)
+    dictionary_entry = factory.SubFactory(DictionaryEntryFactory)
     category = factory.SubFactory(CategoryFactory)
 
 
@@ -65,8 +67,8 @@ class DictionaryEntryRelatedCharacterFactory(DjangoModelFactory):
     class Meta:
         model = dictionary.DictionaryEntryRelatedCharacter
 
-    dictionary_entry = factory.SubFactory(factories.DictionaryEntryFactory)
-    character = factory.SubFactory(factories.CharacterFactory)
+    dictionary_entry = factory.SubFactory(DictionaryEntryFactory)
+    character = factory.SubFactory(CharacterFactory)
 
 
 class PartOfSpeechFactory(DjangoModelFactory):
@@ -74,12 +76,12 @@ class PartOfSpeechFactory(DjangoModelFactory):
         model = PartOfSpeech
 
     title = factory.Sequence(lambda n: "Part of Speech %03d" % n)
-    created_by = factory.SubFactory(factories.UserFactory)
-    last_modified_by = factory.SubFactory(factories.UserFactory)
+    created_by = factory.SubFactory(UserFactory)
+    last_modified_by = factory.SubFactory(UserFactory)
 
 
 class WordOfTheDayFactory(DjangoModelFactory):
-    dictionary_entry = factory.SubFactory(factories.DictionaryEntryFactory)
+    dictionary_entry = factory.SubFactory(DictionaryEntryFactory)
     date = datetime.today()
 
     class Meta:
