+++ firstvoices/backend/tests/factories/media_factories.py
@@ -2,14 +2,14 @@ import factory
 from factory.django import DjangoModelFactory
 
 from backend.models.media import Audio, AudioSpeaker, Image, Person, Video
-from backend.tests import factories
+from backend.tests.factories.access import SiteFactory
 
 
 class ImageFactory(DjangoModelFactory):
     class Meta:
         model = Image
 
-    site = factory.SubFactory(factories.SiteFactory)
+    site = factory.SubFactory(SiteFactory)
     title = factory.Sequence(lambda n: "Image-%03d" % n)
     content = factory.django.ImageField()
 
@@ -18,7 +18,7 @@ class VideoFactory(DjangoModelFactory):
     class Meta:
         model = Video
 
-    site = factory.SubFactory(factories.SiteFactory)
+    site = factory.SubFactory(SiteFactory)
     title = factory.Sequence(lambda n: "Video-%03d" % n)
     content = factory.django.FileField()
 
@@ -27,7 +27,7 @@ class PersonFactory(DjangoModelFactory):
     class Meta:
         model = Person
 
-    site = factory.SubFactory(factories.SiteFactory)
+    site = factory.SubFactory(SiteFactory)
     name = factory.Sequence(lambda n: "Person %03d" % n)
 
 
@@ -35,7 +35,7 @@ class AudioFactory(DjangoModelFactory):
     class Meta:
         model = Audio
 
-    site = factory.SubFactory(factories.SiteFactory)
+    site = factory.SubFactory(SiteFactory)
     title = factory.Sequence(lambda n: "Audio-%03d" % n)
     content = factory.django.FileField()
 
@@ -44,6 +44,6 @@ class AudioSpeakerFactory(DjangoModelFactory):
     class Meta:
         model = AudioSpeaker
 
-    site = factory.SubFactory(factories.SiteFactory)
+    site = factory.SubFactory(SiteFactory)
     audio = factory.SubFactory(AudioFactory)
     speaker = factory.SubFactory(PersonFactory)
